chore: 각인 이미지 가져오는 api 주소 변경 #17
- 캐릭터 정보가 아니라 이미지 url만 가져옴
